submission/supplementary/code/core/basis.py
"""Core implementation of fractal basis functions."""

from typing import Dict, List, Optional, Tuple, Union
from dataclasses import dataclass
import numpy as np
from sympy import (
    Symbol, symbols, exp, integrate, conjugate, sqrt,
    hermite, Expr, diff, oo, I, log, pi
)
import cmath
import warnings
from .physics_constants import (
    ALPHA_VAL, X, E, T, P, Z_MASS,
    GAMMA_1, GAMMA_2, GAMMA_3,
    g1_REF, g2_REF, g3_REF,
    ALPHA_REF, HBAR, C
)
from .transforms import lorentz_boost, gauge_transform
from .types import Energy, FieldConfig, WaveFunction
from .modes import ComputationMode
from .field import UnifiedField
from .errors import PhysicsError, ValidationError

class FractalBasis(UnifiedField):
    """
    Implements recursive fractal basis functions.
    
    Extends UnifiedField to provide specific implementation using fractal basis expansion.
    The basis functions form a complete orthonormal set that diagonalizes the evolution
    operator.
    
    Key features:
    - Recursive construction via scaling relations
    - Built-in energy scale dependence
    - Automatic normalization and error estimation
    """
    
    # Class constants
    STABILITY_THRESHOLD = 1e-10
    MAX_ITERATIONS = 1000
    E0 = Z_MASS  # Z boson mass in GeV
    E0_MIN = 1.0
    E0_MAX = 1000.0
    N_STABLE_MAX = 50
    LOG_NORM_THRESHOLD = 100
    
    def __init__(self, alpha: float = ALPHA_VAL, mode: ComputationMode = ComputationMode.MIXED):
        """Initialize fractal basis with scaling parameter."""
        super().__init__(alpha=alpha, mode=mode)
        self.scaling_dimension = 1.0
    
    # Core implementation methods (overriding UnifiedField)
    def _solve_field_equations(self, config: FieldConfig) -> WaveFunction:
        """Solve field equations using fractal basis expansion."""
        n = config.dimension
        E = config.mass
        
        # Compute light-cone coordinates
        u = (T + X/C)/sqrt(2)  # Retarded time
        v = (T - X/C)/sqrt(2)  # Advanced time
        
        # Scale coordinates properly
        scaled_u = self.alpha**n * u
        scaled_v = self.alpha**n * v
        
        # Get generator function in light-cone coordinates
        F = self._generator_function(scaled_u, scaled_v)
        
        # Apply modulation and scaling
        modulation = self._modulation_factor(n, Energy(E))
        
        # Combine all factors
        psi = self.alpha**n * F * modulation * exp(-I*E*T/HBAR)
        return self.normalize(psi)

    def _compute_evolution_operator(self, energy: Energy) -> WaveFunction:
        """Compute evolution operator in fractal basis."""
        k = energy.value / self.alpha
        # Relativistic dispersion relation
        omega = sqrt(k**2 + (self.alpha*C/HBAR)**2)
        phase = exp(-I * omega * T)
        return phase * self._scaling_operator(k)

    # Basis-specific helper methods
    def _generator_function(self, u: Symbol, v: Symbol) -> WaveFunction:
        """Generate basis function core."""
        # Lorentz invariant combination
        s = u*v  # Spacetime interval
        return exp(-s/2)
    
    def _modulation_factor(self, n: int, E: Energy) -> WaveFunction:
        """Compute energy-dependent modulation."""
        k = E.value / self.alpha**n
        # Relativistic modulation
        p = k/C  # Momentum
        return exp(-p**2 * (X**2 - C**2*T**2)/(2*HBAR**2))

    def _scaling_operator(self, k: float) -> WaveFunction:
        """Compute scaling operator."""
        # Include proper relativistic scaling
        gamma = 1/sqrt(1 - (k/(C*self.alpha))**2)
        return exp(self.scaling_dimension * gamma * log(k))

    # Computation methods
    def compute(self, n: int, E: Energy = Energy(1.0)) -> WaveFunction:
        """Compute nth basis function at energy E."""
        self._validate_inputs(n, E.value)
        config = FieldConfig(mass=E.value, dimension=n, coupling=self.alpha)
        return self._solve_field_equations(config)

    def compute_with_errors(self, n: int, E: float = 1.0) -> Dict[str, Expr]:
        """Compute basis function with error estimates."""
        psi = self.compute(n, Energy(E))
        norm_error = abs(self.check_orthogonality(n, n) - 1.0)
        trunc_error = self.alpha**(n+1) * abs(self._generator_function(X))
        quad_error = 1e-8 * abs(psi)
        return {
            'function': psi,
            'normalization_error': norm_error,
            'truncation_error': trunc_error,
            'integration_error': quad_error,
            'total_error': norm_error + trunc_error + quad_error
        }

    # Remove duplicate methods that are already in UnifiedField:
    # - normalize
    # - compute_inner_product
    # - apply_gauge_transformation
    # - compute_basis_function (replaced by compute)

    def validate_level(self, n: int) -> None:
        """Validate basis function level."""
        if not isinstance(n, int) or n < 0:
            raise PhysicsError("Level n must be non-negative integer")

    def validate_energy(self, E: float) -> None:
        """Validate energy parameter."""
        if not isinstance(E, (int, float)) or E <= 0:
            raise PhysicsError("Energy must be positive")

    def _validate_inputs(self, n: int, E: float) -> None:
        """
        Validate basis function inputs.
        
        Args:
            n: Level number
            E: Energy value
            
        Raises:
            PhysicsError: If inputs are invalid
        """
        if not isinstance(n, int) or n < 0:
            raise PhysicsError("Level n must be non-negative integer")
        if not isinstance(E, (int, float)) or E <= 0:
            raise PhysicsError("Energy must be positive")
